api/requirements.txt
opencv-python==4.7.0.72
flask==2.2.3
gphoto2==2.3.4
pillow
kafka-python==2.0.2
minio
